Added "TravisBuilder" script, added some things to core script

+++ script.php
@@ -2,38 +2,50 @@
 $deployBranch = !getenv("DEPLOY_BRANCH") ? getenv("TRAVIS_BRANCH") : getenv("DEPLOY_BRANCH");
 $token = getenv("TOKEN");
 putenv("TOKEN=''");
-$buildName = getenv("BUILD_NAME");
 $pullRequest = getenv("TRAVIS_PULL_REQUEST") !== false;
+
+$rootPath = rtrim(getenv("TRAVIS_BUILD_DIR"), "/");
+$rootPath = explode("/", $rootPath);
+array_pop($rootPath);
+$rootPath = implode("/", $rootPath) . "/";
+
+$pharPath = $rootPath . "build";
+putenv("PHAR_PATH=$pharPath");
 if($pullRequest){
     echo "[Info] 'Pull Request' detected, build will not be deployed.";
 }else{
-    echo "[Info] '$deployBranch' is the target 'Deploy Branch'";
+    echo "[Info] '$deployBranch' is the target Deploy-Branch";
 }
 if(!$token){
     echo "[Warning] No 'GitHub Token' provided, build will not be deployed.";
 }
-echo "[Info] '" . (!$buildName ? "DevTools" : $buildName) . "' is the assigned 'Build Name' for the file.";
-echo "Setting up environment...";
-/*TODO:
- * Setup environment
- *  - Download and run PocketMine-MP
- *  - Install DevTools
- * Exit script if PocketMine-MP Crashes
- * ...
- */
+echo "[Info] Setting up environment...";
+chdir("$rootPath");
+exec("mkdir server");
+exec("mkdir build");
+chdir("server");
+exec("mkdir plugins");
+exec("cp " . getenv("TRAVIS_BUILD_DIR") . "/travis/TravisBuilder.php " . $rootPath . "server/plugins");
+exec("cp " . getenv("TRAVIS_BUILD_DIR") . " " . $rootPath . "server/plugins/" . array_pop(explode("/", getenv("TRAVIS_REPO_SLUG"))));
+exec("curl -sL get.pocketmine.net | bash -s - -v " . (getenv("PM_VERSION") !== false ? getenv("PM_VERSION") : "stable"));
+$pmLine = null;
+echo "[Info] Starting PocketMine-MP...";
+exec("./start.sh --diable-ansi --no-wizard", $t, $pmLine);
 
-if(!$pullRequest && $token !== false){
-    echo "Preparing to deploy...";
-    // Move into 'DevTools' directory...
+// TODO: Exit state of PMMP
+
+if(is_dir($pharPath) && !$pullRequest && $token !== false){
+    echo "[Info] Preparing to deploy...";
+    chdir("$pharPath");
     exec("git init");
-    exec("git remote add origin https://$TOKEN@github.com/" . getenv("TRAVIS_REPO_SLUG" . ".git"));
+    exec("git remote add origin https://$TOKEN@github.com/" . getenv("TRAVIS_REPO_SLUG"));
     exec("git fetch origin");
-    exec("git config user.name \"LegendsOfMCPE-Bot\"");
-    exec("git config user.email \"LoM_Bot@travis.ci\"");
-    exec("git add -A");
-    echo "Creating commit...";
+    exec("git config user.name \"iksaku's BuilderBot\"");
+    exec("git config user.email \"iksaku_Bot@travis.ci\"");
+    exec("git add .");
+    echo "[Info] Creating commit...";
     exec("git commit -m \"New Build! Revision: " . getenv("TRAVIS_COMMIT") . "\"");
-    echo "Pushing commit...";
+    echo "[Info] Pushing commit...";
     exec("git push --force --quiet origin HEAD:$deployBranch");
-    echo "Build successfully uploaded!";
+    echo "[Info] Build successfully uploaded!";
 }
